+++ Makefile
@@ -1,0 +1,6 @@
+# Top level makefile
+
+default: LISPB
+
+.DEFAULT:
+	cd src && $(MAKE) $@
