--- fungsimain.c
@@ -1,12 +1,0 @@
-#include <stdio.h>
-#include "fungsi.h"
-
-int main()
-{
-	const char *Input = "mod";
-	int Tipe;
-
-	printf("%d", IsFUNGSI(Input));
-
-	return 0;
-}
