--- test.txt
@@ -1,5 +1,0 @@
-( cdr ( list 3 4 5) ) #
-( + ( + 4 5) (+ 5 6))#
-(setq LL (list 3 4 5 6 7))#
-(cons (car LL) (list 1 2))#
-(reverse LL)#
