--- test1.txt
@@ -1,4 +1,0 @@
-
-(+ 2 3)#
-(save test1.txt)#
-(+ 4 5)#st1.txt)#
